% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JTR.jtr}
\alias{JTR.jtr}
\title{JTR-data}
\format{A data frame with 2941 rows and 19 variables:
\describe{
  \item{TournamentID}{unique tournament key}
  \item{TournamentName}{Name (not always unique)}
  \item{TournamentStart}{announced starting timepoint}
  \item{TournamentEnd}{announced end timepoint}
  \item{TournamentCountry}{country name}
  \item{TournamentCity}{city name}
  \item{TournamentPostalCode}{postal code (as text)}
  \item{TournamentStreet}{street (text)}
  \item{TournamentPlace}{specific name of the location}
  \item{TournamentLatitude}{geographic latitude of the tournament}
  \item{TournamentLongitude}{geographic longitude of the tournament}
  \item{maxParticipants}{maximal allowed number of participants (0 means no limit)}
  \item{nParticipants}{actual number of participants}
  \item{TeamID}{unique team key}
  \item{TeamName}{Name of the team}
  \item{TeamCountry}{country name}
  \item{TeamCity}{city name}
  \item{TeamLatitude}{geographic latitude of the team (coutnry, city)}
  \item{TeamLongitude}{geographic longitude of the team (coutnry, city)}
  \item{nParticipations}{number of registered torunament participation of each team}
  \item{Rank}{rank of the team in the tournament}
}}
\source{
\url{http://turniere.jugger.org/}
}
\usage{
JTR.jtr
}
\description{
A data set of jugger tournament participation
and results over a long time preiode (2009-2017).
}
\keyword{datasets}
