% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JTR.Teams}
\alias{JTR.Teams}
\title{JTR-team data}
\format{A data frame with 474 rows and 7 variables:
\describe{
  \item{TeamID}{unique team key}
  \item{TeamName}{Name of the team}
  \item{TeamCountry}{country name}
  \item{TeamCity}{city name}
  \item{TeamLatitude}{geographic latitude of the team (coutnry, city)}
  \item{TeamLongitude}{geographic longitude of the team (coutnry, city)}
  \item{nParticipations}{number of registered torunament participation of each team}
}}
\source{
\url{http://turniere.jugger.org/}
}
\usage{
JTR.Teams
}
\description{
A data set of jugger teams registered in the JTR
from a long time preiode (2009-2017).
}
\keyword{datasets}
