% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JTR.Tournaments}
\alias{JTR.Tournaments}
\title{JTR-tournament data}
\format{A data frame with 217 rows and 13 variables:
\describe{
  \item{TournamentID}{unique tournament key}
  \item{TournamentName}{Name (not always unique)}
  \item{TournamentStart}{announced starting timepoint}
  \item{TournamentEnd}{announced end timepoint}
  \item{TournamentCountry}{country name}
  \item{TournamentCity}{city name}
  \item{TournamentPostalCode}{postal code (as text)}
  \item{TournamentStreet}{street (text)}
  \item{TournamentPlace}{specific name of the location}
  \item{TournamentLatitude}{geographic latitude of the tournament}
  \item{TournamentLongitude}{geographic longitude of the tournament}
  \item{maxParticipants}{maximal allowed number of participants (0 means no limit)}
  \item{nParticipants}{actual number of participants}
}}
\source{
\url{http://turniere.jugger.org/}
}
\usage{
JTR.Tournaments
}
\description{
A data set of jugger tournaments over a long time preiode (2009-2017).
}
\keyword{datasets}
