% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWISS.R
\name{read_SWISS_tournament}
\alias{read_SWISS_tournament}
\title{Read files of the SWISS tournament software}
\usage{
read_SWISS_tournament(filename)
}
\arguments{
\item{filename}{name of the XML file returned by the program}
}
\value{
a list with the following entries:
\itemize{
\item tournament_version (integer)
\item score_calculation (character)
\item ranking_comparator (character)
\item teams (a data.frame of teams)
  \itemize{
    \item entry_id (integer)
    \item team_name (character)
    \item team_city (character)
  }
\item rounds (a data.frame of all rounds and matches)
  \itemize{
    \item round (integer)
    \item game (interger)
    \item entry_id1 (integer)
    \item entry_id2 (integer)
    \item points1 (integer)
    \item points2 (integer)
    \item finished (logical)
  }
}
}
\description{
\code{read_SWISS_tournament} returns all entries of a SWISS tournament.
}
\details{
The SWISS tournament software is a old (2008) piece of software that allows
to manage a jugger tournament as a variant of the SWISS system. It returns a
XML file with all tournament information, teams, rounds, matches and results.
This function reads those information into R for analysis purposes.
}
\examples{

\dontrun{
hanseaticjuggercup2014 <- read_SWISS_tournament("HanseaticJuggerCupR6")
}
}
\author{
Christian Beck
}
